% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add one number to another number}
\usage{
add(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{Another number}
}
\value{
The result of \code{a} added to \code{b}
}
\description{
Add one number to another number
}
\examples{
add(7,1)
}
