% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_data.R
\docType{data}
\name{dog_data}
\alias{dog_data}
\title{Number of spots on certain dogs and breeds}
\format{A data frame with 4 rows and 6 variables:
\describe{
     \item{Dog_name}{Name of the dog}
     \item{breed}{the breed of the dog}
     \item{age}{the age in years of the dog}
     \item{num_spots}{the number of spots on the dog's coat}
     }}
\usage{
dog_data
}
\description{
These data were collected during a trip to a dog park in Austin full of adorable pups.
}
\keyword{datasets}
