% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract.R
\name{subtract}
\alias{subtract}
\title{Subtract one number from another number}
\usage{
subtract(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{Another number}
}
\value{
The result of \code{b} subtracted from \code{a}
}
\description{
Subtract one number from another number
}
\examples{
subtract(7,1)
}
